chore(settings) Move language to general tab (#13405)
