+++ react/features/settings/functions.any.ts
@@ -114,17 +114,22 @@ export function getNotificationsMap(stateful: IStateful) {
 export function getMoreTabProps(stateful: IStateful) {
     const state = toState(stateful);
     const stageFilmstripEnabled = isStageFilmstripEnabled(state);
+    const language = i18next.language || DEFAULT_LANGUAGE;
+    const configuredTabs: string[] = interfaceConfig.SETTINGS_SECTIONS || [];
 
     // when self view is controlled by the config we hide the settings
     const { disableSelfView, disableSelfViewSettings } = state['features/base/config'];
 
     return {
+        currentLanguage: language,
         disableHideSelfView: disableSelfViewSettings || disableSelfView,
         hideSelfView: getHideSelfView(state),
         iAmVisitor: iAmVisitor(state),
+        languages: LANGUAGES,
+        maxStageParticipants: state['features/base/settings'].maxStageParticipants,
+        showLanguageSettings: configuredTabs.includes('language'),
         showPrejoinPage: !state['features/base/settings'].userSelectedSkipPrejoin,
         showPrejoinSettings: state['features/base/config'].prejoinConfig?.enabled,
-        maxStageParticipants: state['features/base/settings'].maxStageParticipants,
         stageFilmstripEnabled
     };
 }
@@ -195,20 +200,15 @@ export function getProfileTabProps(stateful: IStateful) {
     } = state['features/base/conference'];
     const { hideEmailInSettings } = state['features/base/config'];
     const localParticipant = getLocalParticipant(state);
-    const language = i18next.language || DEFAULT_LANGUAGE;
-    const configuredTabs: string[] = interfaceConfig.SETTINGS_SECTIONS || [];
 
     return {
         authEnabled: Boolean(conference && authEnabled),
         authLogin,
-        currentLanguage: language,
         displayName: localParticipant?.name,
         email: localParticipant?.email,
         hideEmailInSettings,
         id: localParticipant?.id,
-        languages: LANGUAGES,
-        readOnlyName: isNameReadOnly(state),
-        showLanguageSettings: configuredTabs.includes('language')
+        readOnlyName: isNameReadOnly(state)
     };
 }
